+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Added options to read and stat files immediately after creation while they are still open. (See "--readinline" and "--statinline".)
 * Added option to sleep between test phases. (See "--phasedelay".)
 * Added option to rotate hosts list for service instances between phases to avoid caching effects. (See "--rotatehosts".)
-* Added support for S3 object and bucket ACL PUT/GET. (See "--s3aclput", "--s3baclput".)
+* Added support for S3 object and bucket ACL PUT/GET. (See "--s3aclput", "--s3aclputinl", "--s3baclput".)
 * Added new ops log to log all IO operations (open, read, ...). (See "--opslog").
 
 ### General Changes